backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
PyMuPDF==1.24.10
python-docx==1.1.2
Pillow==10.4.0
torch==2.4.1
open-clip-torch==2.26.1
torchaudio==2.4.1
rich==13.8.1

aiohttp==3.10.5
PyYAML==6.0.2
# Optional LLM backends (install as needed):
# gpt4all
llama-cpp-python==0.2.90
# open-clip-torch already included for image embeddings
reportlab==4.2.2


